HBASE-26641 Split TestMasterFailoverWithProcedures (#3994)
